refactor: Rename AssigneeSelector and update import paths
- Renamed AssigneeSelector component to follow kebab-case naming convention
- Updated import paths across multiple components
- Removed unused props and simplified component imports
- Triggered task groups refresh on filter changes
